Add default callback for invisible recaptcha #220
This adds a default callback that calls the function
`invisibleRecaptchaSubmit` if no callback was passed to
`invisible_recaptcha_tags`. It will traverse the DOM upwards to find
the closest form to submit.

The drawback to this approach is that it won't work if there are
multiple recaptcha forms. If you want to use invisible_recaptcha_tags
with multiple forms, you would still need to use custom callbacks.