Move documentation for the Recaptcha initialiser
Have moved the section regarding the Recaptcha gem's initialiser closer
to the rest of the Rails configuration instructions in the README as
this was originally quite hard to find. I have also fixed a typo in that
code as the gem only works if a `Recaptcha.configuration` block has been
defined as opposed to a `Recaptcha.configure` block.